Split the ast module into separate files
The file `ast.rs` has become just too big. And since I've just learned
you don't need to put every `impl X` into the module where `X` is
defined, let's put that into action.
